+++ .build/sketch/ReadSensorData.h
@@ -13,7 +13,7 @@
 #define READSENSORDATA_H
 
 void SetupSensors();
-float Round(float);
+double Round(float);
 float ReadHumidity();
 float ReadTemperature();
 float ReadPressure();
